Allow independant nested projects
I want unknown to be able to support "subprojects" - projects that are
nested inside the directory of another project.  It sort of did this
already, but any changes to files inside a nested project would also be
considered a change of the parent projects files.

This updates projects to contain a set of "excluded paths" and updates
the changed command to take these paths into account.

Will also need to update the run command to do the same, which I'll do
in a future commit.